add 3 account cannot delete

+++ src/users/users.service.ts
@@ -152,7 +152,19 @@ export class UsersService {
       throw new BadRequestException('Not a valid ObjectId!');
     }
 
-    return await this.userModel.delete({ _id: id }, deletedBy);
+    const user = await this.userModel.findById(id);
+    if (user) {
+      if (
+        user.email === 'admin@gmail.com' ||
+        user.email === 'user@gmail.com' ||
+        user.email === 'guest@gmail.com'
+      ) {
+        throw new BadRequestException(
+          'Định mệnh, xóa tài khoản này lấy gì mà test @@',
+        );
+      }
+      return await this.userModel.delete({ _id: id }, deletedBy);
+    }
   }
 
   async register(registerUserDto: RegisterUserDto) {
